Remove uses of protoc-gen-go/generator
Replace with internal copy of CamelCase
