+++ examples/internal/server/BUILD.bazel
@@ -19,10 +19,10 @@ go_library(
         "//examples/internal/proto/examplepb:go_default_library",
         "//examples/internal/proto/sub:go_default_library",
         "//examples/internal/proto/sub2:go_default_library",
+        "//internal/casing:go_default_library",
         "//runtime:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_golang_protobuf//protoc-gen-go/generator:go_default_library_gen",
         "@com_github_rogpeppe_fastuuid//:go_default_library",
         "@go_googleapis//google/rpc:errdetails_go_proto",
         "@io_bazel_rules_go//proto/wkt:duration_go_proto",
