+++ internal/casing/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["camel.go"],
+    importpath = "github.com/grpc-ecosystem/grpc-gateway/internal/casing",
+    visibility = ["//:__subpackages__"],
+)
