+++ protoc-gen-grpc-gateway/internal/gengateway/BUILD.bazel
@@ -11,12 +11,12 @@ go_library(
     ],
     importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway/internal/gengateway",
     deps = [
+        "//internal/casing:go_default_library",
         "//protoc-gen-grpc-gateway/descriptor:go_default_library",
         "//protoc-gen-grpc-gateway/generator:go_default_library",
         "//utilities:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
-        "@com_github_golang_protobuf//protoc-gen-go/generator:go_default_library_gen",
         "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
     ],
 )
