+++ protoc-gen-swagger/genswagger/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
     importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-swagger/genswagger",
     deps = [
         "//internal:go_default_library",
+        "//internal/casing:go_default_library",
         "//protoc-gen-grpc-gateway/descriptor:go_default_library",
         "//protoc-gen-grpc-gateway/generator:go_default_library",
         "//protoc-gen-swagger/options:go_default_library",
